Fix/clippy

+++ generator/templates/audio_param.tmpl.rs
@@ -80,7 +80,7 @@ fn linear_ramp_to_value_at_time(ctx: CallContext) -> Result<JsUndefined> {
     let obj = napi_obj.unwrap();
 
     let value = ctx.get::<JsNumber>(0)?.get_double()? as f32;
-    let end_time = ctx.get::<JsNumber>(1)?.get_double()? as f64;
+    let end_time = ctx.get::<JsNumber>(1)?.get_double()?;
     obj.linear_ramp_to_value_at_time(value, end_time);
 
     ctx.env.get_undefined()
@@ -93,7 +93,7 @@ fn exponential_ramp_to_value_at_time(ctx: CallContext) -> Result<JsUndefined> {
     let obj = napi_obj.unwrap();
 
     let value = ctx.get::<JsNumber>(0)?.get_double()? as f32;
-    let end_time = ctx.get::<JsNumber>(1)?.get_double()? as f64;
+    let end_time = ctx.get::<JsNumber>(1)?.get_double()?;
     obj.exponential_ramp_to_value_at_time(value, end_time);
 
     ctx.env.get_undefined()
@@ -108,8 +108,8 @@ fn set_value_curve_at_time(ctx: CallContext) -> Result<JsUndefined> {
     let mut typed_array_values = ctx.get::<JsTypedArray>(0)?.into_value()?;
     let values: &mut [f32] = typed_array_values.as_mut();
 
-    let start_time = ctx.get::<JsNumber>(1)?.get_double()? as f64;
-    let duration = ctx.get::<JsNumber>(2)?.get_double()? as f64;
+    let start_time = ctx.get::<JsNumber>(1)?.get_double()?;
+    let duration = ctx.get::<JsNumber>(2)?.get_double()?;
     obj.set_value_curve_at_time(values, start_time, duration);
 
     ctx.env.get_undefined()
@@ -122,8 +122,8 @@ fn set_target_at_time(ctx: CallContext) -> Result<JsUndefined> {
     let obj = napi_obj.unwrap();
 
     let value = ctx.get::<JsNumber>(0)?.get_double()? as f32;
-    let start_time = ctx.get::<JsNumber>(1)?.get_double()? as f64;
-    let time_constant = ctx.get::<JsNumber>(2)?.get_double()? as f64;
+    let start_time = ctx.get::<JsNumber>(1)?.get_double()?;
+    let time_constant = ctx.get::<JsNumber>(2)?.get_double()?;
     obj.set_target_at_time(value, start_time, time_constant);
 
     ctx.env.get_undefined()
@@ -135,7 +135,7 @@ fn cancel_scheduled_values(ctx: CallContext) -> Result<JsUndefined> {
     let napi_obj = ctx.env.unwrap::<NapiAudioParam>(&js_this)?;
     let obj = napi_obj.unwrap();
 
-    let cancel_time = ctx.get::<JsNumber>(0)?.get_double()? as f64;
+    let cancel_time = ctx.get::<JsNumber>(0)?.get_double()?;
     obj.cancel_scheduled_values(cancel_time);
 
     ctx.env.get_undefined()
@@ -147,7 +147,7 @@ fn cancel_and_hold_at_time(ctx: CallContext) -> Result<JsUndefined> {
     let napi_obj = ctx.env.unwrap::<NapiAudioParam>(&js_this)?;
     let obj = napi_obj.unwrap();
 
-    let cancel_time = ctx.get::<JsNumber>(0)?.get_double()? as f64;
+    let cancel_time = ctx.get::<JsNumber>(0)?.get_double()?;
     obj.cancel_and_hold_at_time(cancel_time);
 
     ctx.env.get_undefined()
